Appearance Improvements of Thomas Hübner
- 3 New settings in "Appearance" section
--- row height
--- colorize yes/no
--- Threshold for colorize heavy used channels

If a row height < 35 is given by user the list switches to the slim
design with num user in front of channelname and limit to one line. If
row height >=35 the legacy layout is used. Values are corrected to be
min 30 and max 50 (no clue whether a min/max limit is possible in the
preferences XML file - if yes it should be used there)

Colorize ON/OFF is self explaining as well as the threshold. I have
found no possibility to inactivate (grey out) the threshold setting if
colorize is NO. This should be done if possible